+++ src/db/guitar/main.js
@@ -1,5 +1,6 @@
 export default {
   strings: 6,
+  fretsOnChord: 5,
   keys: [ 'C', 'C#', 'D', 'Eb', 'E', 'F', 'F#', 'G', 'Ab', 'A', 'Bb', 'B' ],
   tunnings: {
     standard: [ 'E', 'B', 'G', 'D', 'A', 'E' ]