+++ src/db/ukelele/chords/C/7.js
@@ -1,0 +1,7 @@
+export default {
+  key: 'C',
+  suffix: '7',
+  positions: [{
+    frets: '1000'
+  }]
+}