+++ src/db/ukelele/chords/C/index.js
@@ -1,5 +1,11 @@
 import major from './major'
+import minor from './minor'
+import seven from './7'
+import min7 from './min7'
 
 export default [
-  major
+  major,
+  minor,
+  seven,
+  min7
 ]