+++ src/db/ukelele/chords/C/min7.js
@@ -1,0 +1,8 @@
+export default {
+  key: 'C',
+  suffix: 'min7',
+  positions: [{
+    frets: '3333',
+    barres: 3
+  }]
+}