+++ src/db/ukelele/chords/C/minor.js
@@ -1,0 +1,7 @@
+export default {
+  key: 'C',
+  suffix: 'minor',
+  positions: [{
+    frets: '3330'
+  }]
+}