+++ src/db/ukelele/main.js
@@ -1,5 +1,6 @@
 export default {
   strings: 4,
+  fretsOnChord: 5,
   name: 'soprano',
   keys: [ 'C', 'C#', 'D', 'Eb', 'E', 'F', 'F#', 'G', 'Ab', 'A', 'Bb', 'B' ],
   tunnings: {